requirements.txt
streamlit
python-dotenv
groq
gtts
elevenlabs
speechrecognition
pydub
ffmpeg-python
deep-translator
